BagFile.jl

BagFile.jl is a Julia package for parsing ROS bag files

The aim is to provide a library for working with ROS bags in the same way as the command line tool from ROS.

BagFile.jl

BagFile.BagFile — Module

BagFile

this module reproduces the "rosbag command-line tool" for Julia record Record a bag file with the contents of specified topics.

info IMPLEMENTED Summarize the contents of a bag file.

play (not implemented) Play back the contents of one or more bag files.

check (not implemented) Determine whether a bag is playable in the current system, or if it can be migrated.

fix (not implemented) Repair the messages in a bag file so that it can be played in the current system.

filter (not implemented) Convert a bag file using Python expressions.

compress (not implemented) Compress one or more bag files.

decompress (not implemented) Decompress one or more bag files.

reindex (not implemented) Reindex one or more broken bag files.

.

BagFile.BagFileData — Type

BagFileData

struct con metadata del BagFile Para el constructor solo se necesita el path y luego con la funcion open se populan los datos es posible que se agregue el IO y un puntero de referncia?

BagFile.ROSMessages — Module

ROSMessages

This module contains the structs and constructors for all ROS serialized data

BagFile.leer_bag_header — Function

leer_bag_header

Funcion para leer la primera linea del rosbag y verificar que esta OK

BagFile.leer_record — Function

leer_record
funcion que lee un record, avanza el io hasta el final del record y devuelve un struct del tipo Record

leer_record

funcion que lee un record, en la posiscion por del chunk

BagFile.topics — Function

topics

Funcion que recibe un bag y devuelve lista (o dict) con los topics

BagFile.parse_connection — Function

parse_connection

function that receives a connection record and returns a Dict with the data

BagFile.parse_chunk_info — Function

parse_chunk_info

function that receives a chunk info record and returns a Dict with the data

BagFile.OpenBag — Function

OpenBag

Funcion que inicialize un BagFileData y devuelve el struct BagFileData con la metadata